perf: Pack prize index (#12)
* perf: Pack prize index

* perf: Pack entries count/index

* perf: Pack minimum profit basis points

* perf: Pack winners count / cumulative winners count

* perf: Pack cutoffTime and drawnAt

* refactor: Rename pricingIndex

+++ test/foundry/Raffle.claimFees.t.sol
@@ -34,7 +34,7 @@ contract Raffle_ClaimFees_Test is TestHelpers {
 
         vm.startPrank(user1);
         looksRareRaffle.createRaffle({
-            cutoffTime: block.timestamp + 86_400,
+            cutoffTime: uint40(block.timestamp + 86_400),
             minimumEntries: 107,
             maximumEntries: 200,
             maximumEntriesPerParticipant: 100,
